requirements.txt
# Healthcare Regulatory Intelligence - Python Dependencies
# Generated: 2026-04-19

# Web Framework
streamlit>=1.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Databricks SDK & Integrations
databricks-sdk>=0.18.0
databricks-sql-connector>=3.0.0
databricks-vectorsearch>=0.22

# ML & Embeddings
mlflow>=2.10.0

# LangGraph & LangChain
langgraph>=0.0.26
langchain>=0.1.0
langchain-community>=0.0.20

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
